+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -11,6 +11,7 @@
 #### Other Changes
 * Changed to use `PartitionKeyRangeCache` to get partition key range during startup and split handling. - [46700](https://github.com/Azure/azure-sdk-for-java/pull/46700)
 * Changed to use lower casing http header names for gateway response. - [46736](https://github.com/Azure/azure-sdk-for-java/pull/46736)
+* Changed timestamp format to be consistent in leases for CFP. - []()
 
 ### 4.74.0 (2025-09-05)